Add mech to display a comboButton when a combo can be formed
At time of writing, it has been discovered that Unity has this UI kit,
which is really handy when constructing a UI. Currently the Combo
displaying on screen, as constructed in commit
41e4664d1c129296327d63e95ee7308e4eded684, is constructing new GameObjects
in Unity. This is unclear, since those combo displayer elements are UI,
not part of the game. Therefore, they should be moved to using Unity.UI
instead.

Key Points:
 - Use Unity.UI and Canvas to display Combo Elements instead of
   constructing new GameObjects.

See Also:
 # Combo sequence displayer commit:
     41e4664d1c129296327d63e95ee7308e4eded684