Enable switch back to CancelTile state if battle is unresolved
In the process of implementing this feature, many flaws with the current
system has been exposed

1. A lot of tight coupling between controllers have been introduced. For
instance, BattleController hard-references GameController to tell
GameController to change states when battle finishes. This is a stupid
design choice, and should be avoided at all costs. But for now, due to
time concerns, there are no time to introduce major refactors of
structures, so for now it will remain like this.
Possible solutions:
 - Use StrangeIoC, the framework introduces the possibility of using
   Dependency Injection and Signal&Slots, which eliminates hard coupling.

2. Getting references to other objects are becoming a major problem.
Currently there are two naive approaches in Unity:
 i.	Use GameObject.Find(GameObjectName).GetComponent<scriptName>();
 ii.	Use Drag & drop feature of Unity
The Find method is extremely costly, therefore not recommended. The drag
and drop approach is too delicate, as anyone in the UnityEditor can modify
that, and it shouldn't be this way. Both method creates tight coupling and
spaghetti code.

3. Now I am debating on whether I should really have Board, BoardDisplay
and BoardController. A lot of resetting board code are crude, I have not
yet looked into possible side effects. At time of writing those code, no
effort has been made to accommodate for resetting board.

At this point, it is safe to conclude that this project currently has a
very bad architecture. When time permits, a thorough investigation and
refactor to use DI, Signal & Slot method is strongly advised.