Fix a bug with BattleResolving not shifting
This bug only occurs when the player cancels 0 or 1 tile. The problem was
that with 0 or 1 tile, the distance to move isn't being initialized. Fixed
the issue by directly assign the battleResolveContainer's starting
position as the starting point, and the distance to move is simply
calculated from using the assumed widthSegment. Where these points are
doesn't really matter, all we need is the distance.