Let ComboModel determine combo using info from comboTree
ComboModel used to have a defined int that is used to determine the number
of tiles that should appear on the screen, and the model uses this value
in order to determine how far should it traverse comboTree to look for
combos. This is a messy design, so this commit let ComboTree to know the
longest combo length, and ComboModel will grab this info instead.
The number of tiles that appears on screen obviously belongs to ComboView